+++ CHANGELOG.md
@@ -8,6 +8,16 @@
 
 ## Bug Fixes
 - [#758](https://github.com/helmholtz-analytics/heat/pull/758) Fix indexing inconsistencies in `DNDarray.__getitem__()`
+### Linear Algebra
+- [#718](https://github.com/helmholtz-analytics/heat/pull/718) New feature: `trace()`
+- [#768](https://github.com/helmholtz-analytics/heat/pull/768) New feature: unary positive and negative operations
+### Misc.
+- [#761](https://github.com/helmholtz-analytics/heat/pull/761) New feature: `result_type`
+
+## Breaking Changes
+
+## Bug fixes
+- [#768](https://github.com/helmholtz-analytics/heat/pull/768) Fixed an issue where `deg2rad` and `rad2deg`are not working with the 'out' parameter.
 
 
 # v1.0.0
@@ -38,10 +48,13 @@
 
 ### Factories
 - [#707](https://github.com/helmholtz-analytics/heat/pull/707) New feature: `asarray()`
+
 ### I/O
 - [#559](https://github.com/helmholtz-analytics/heat/pull/559) Enhancement: `save_netcdf` allows naming dimensions, creating unlimited dimensions, using existing dimensions and variables, slicing
+
 ### Linear Algebra
 - [#658](https://github.com/helmholtz-analytics/heat/pull/658) Bugfix: `matmul` on GPU will cast away from `int`s to `float`s for the operation and cast back upon its completion. This may result in numerical inaccuracies for very large `int64` DNDarrays
+
 ### Logical
 - [#711](https://github.com/helmholtz-analytics/heat/pull/711) `isfinite()`, `isinf()`, `isnan()`
 - [#743](https://github.com/helmholtz-analytics/heat/pull/743) `isneginf()`, `isposinf()`